requirements/dev.txt
# Inclui as dependências do base.txt
-r ./base.txt

black==23.3.0
flake8==6.1.0
pytest==7.2.0
pytest-django==4.5.2
pytest-cov==3.0.0
django-extensions==3.2.0
django-debug-toolbar==3.3.0
ipython==8.7.0
mypy==0.982
